// frontend/src/generated/platform.ts
export enum PlatformTypeEnum {
  UNKNOWN_PLATFORM = 0,
  GROUND = 1,
  SURFACE = 2,
  AIR = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface Platform {
  Id: PlatformId | undefined;
  name: string;
  type: PlatformTypeEnum;
  maxSpeed: number;
  maxZ: number;
  friendly: boolean;
  color: string;
  reportingFrequency: number;
  readonly: boolean;
  waypoints: Waypoint[];
}

export interface Waypoint {
  Id: WaypointId | undefined;
  index: number;
  lat: number;
  lon: number;
  z: number;
  speedKts: number;
  smaj: number;
  smin: number;
  orientation: number;
  datetimeUTC: string;
}